Threading Support
It is now possible to request for work to be done in the background,
making use of the parallel processing resources on a machine

Channels provide a thread safe communication, which can be used to
queue up results from one or more provider threads.  One or more
subscribers may then process the results.

Work is managed by a Task_Master singleton who hold responsibility for
allocating runn time, and persisting / clearning up resources.

The mechanism is largely abstracted through a GO macro (see Tests/Threads)

Logging has been made thread-safe through a mutex. As a result it should
be disabled for non-testing builds, as it will tie the system (by in
large) to a single execution stream. Checking LOG_ENABLED is atmoic, but
does not require a mutex lock.