Replace CEF with custom WebSockets Server
=== Overview ===
Starting with @f2c054f158e84f33df690bccf3bf965a1e36aecd, libCEF provided
a wrapper around the chromium embedded framework for building web
applications with a C++ backend. As a C++ framework libCEF was
successful in its goal abstracting the CEF interface sufficiently to
develop modular applications, such as gcgv (see
Grauniad/Call-Graph-Viewer).

CEF was originally chosen as it allowed building the front end using
HTML/JS/CSS (ideal for leveraging he large number of web-based
visualisation libraries), whilst providing an interface to the C++
back-end to do complex logic. Whilst this proved to be very effective
in building the application, two key disadvantages were found with CEF:

   1. DLL Hell. CEF requires all the libraries used by a modern Chromium
      install, and can be very burdensome on "stable" Linux distros,
      such as RHEL. While it is possible to custom build Chrome and CEF
      for these platforms, this is a tedious process.
   2. Debugging is hard. CEF bundles your application logic into the
      heart of a fully-featured chromium browser, and it is only
      accessed via call-backs from the event loop. These call-backs
      happen across a variety of sub-processes and threads, making
      debugging and profiling needlessly difficult. Of particular
      annoyance is the fact that it is not possible to valgrind-memcheck
      a CEF process.

The goal is to create a new library providing the same abstraction
between a web-based client and C++ server, but which is easier to
deploy, and has better debugging tools.

=== WebSockets ===
Under CEF, client-sever communication was handled by a custom JS
function which could send JSON requests from the client to the server,
and a response back again. With CEF removed, the client must now be a
static web-application and must communicate with the back-end "over the
network". Web-sockets is a simple network protocol supported by
JS which allows this communication without us having to deploy a full
web-server.

A new library , libWebSockets, is provided which can create a very
simple server listening on a custom port. The server will accept new
websocket connections, and can service JSON requests. These requests can
be installed using a RequestReply interface analogous to the one
previously provided by libCEF. This allows existing applications to be
simply migrated.

A third-party library websocketpp, itself a wrapper around
boost::asyncio, is used to handle the websocket protocol and manage
sessions. This is entirely header based and the only run-time dependency
is -lboost-system. The resulting binary is a vanilla, single-threaded,
C++ application and can be developed / deployed in the standard manner.
The front end can be started up by pointing the system browser at the
start page.

=== libJSON Enhancements ===
Previously, the CEF interface presented to the client, was tested using
the CEF test application executing Java-script. With the removal CEF
this was no longer possible and all tests have been migrated into pure
C++ tests. For this to work it is required to process more complex JSON
objects to validate their content on the response. The libJSON library
has been enhanced such that:
   1. Introduced support for embedded JSON objects. It is now possible
      to build up a SimpleParsedJSON object from sub-objects. For
      example:
          {
              "Object" : {
                  "Field1": "Hello World!",
                  "Field2": "",
              }
          }
      can be represented:
          NewStringField(Field1)
          NewStringField(Field2)
          typedef SimpleParsedJSON<Field1,Field2> JSONObject;
          NewEmbededObject(Object,JSONObject);
          typedef SimpleParsedJSON<Object> JSON;
      There is no limit on the level of embeddedness. (An embedded
      objects may itself be made of component objects, and so on).
   2. Introduced support for arrays of anonymous objects. Again these
      objects may themselves by comprised of embedded objects or arrays.
          {
              "Objects": [{
                  "Field1": "Hello World!",
                  "Field2": ""
              }, {
                  "Field1": "Another string",
                  "Field2": "Not a blank string"
              }]
          }
      May be represented as:
          NewStringField(Field1)
          NewStringField(Field2)
          typedef SimpleParsedJSON<Field1,Field2> JSONObject;
          NewObjectArray(Objects,JSONObject)
          typedef SimpleParsedJSON<Objects> JSON;