HTTP requests must now be manually started
Previously the HTTP request object would trigger the request upon construction.
When used as a base class in a multi-threaded application this could cause a
race condition where the completed call-back occurred before construction was
complete on the main thread.

This has been fixed by providing a method to allow derived classes to start
the request once they are ready to handle the response.